requirements.txt
amqp==2.1.4
appdirs==1.4.3
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
cairocffi==0.8.0
CairoSVG==1.0.22
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
cssselect==1.0.1
Django==1.11.1
django-appconf==1.0.2
django-filter==1.0.4
django-pagination-bootstrap==1.2.4
django-ttag==3.0
django-widget-tweaks==1.4.1
flower==0.9.1
futures==3.1.1
html5lib==0.999999999
kombu==4.0.2
lxml==3.7.3
olefile==0.44
packaging==16.8
Pillow==4.1.1
pycparser==2.17
pyparsing==2.2.0
Pyphen==0.9.4
pytz==2017.2
six==1.10.0
tinycss==0.4
tornado==4.2
URLObject==2.4.3
vine==1.1.3
WeasyPrint==0.36
webencodings==0.5.1
webstack-django-sorting==1.0.1